Add go Unit Tests
